Separate utils from getTicks
